+++ lib/cog_api/client.ex
@@ -7,4 +7,5 @@ defmodule CogApi.Client do
   @callback role_index(%Endpoint{}) :: {atom, [%Role{}]}
   @callback role_show(%Endpoint{}, String.t) :: {atom, %Role{}}
   @callback role_create(%Endpoint{}, %{}) :: {atom, %Role{}}
+  @callback role_update(%Endpoint{}, String.t, %{}) :: {atom, %Role{}}
 end